Involved people show up on the edit referral page
For now, passing the involved peopl as data attribute, since the rest of the page is rendered server side
and there is no API to get the involved people for a referral.

I tried using an open source library as the glue code: https://github.com/DeloitteDigitalAPAC/react-habitat
which basically glues together your dom element with a specific component that you can specify as a data attribute(simliar concept
to what is done here).

However there is a bug in this library, and I have opened a github issue: https://github.com/DeloitteDigitalAPAC/react-habitat/issues/3

[#128067971]